Add files via upload
This file contains the environment and actor we worked with throughout the latter half of course. We had to learn about training the actor which was done using reward methods. Through Q-Learning we were able to further refine the actor's choices until not only was it able to complete the maze, but it was able to find more efficient ways of clearing the maze. 

Computer scientists are constantly developing solutions to the various problems they are presented with, through finding solutions to these problems the CS field has been able to enrich the quality of life humanity has to experience. It is important to remember that, as with any tool, technology can help or harm humanity; it is up to those who wield it to decide how technology developed in the CS field is going to affect humanity.